pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "endee-ai-project"
version = "0.1.0"
description = "AI Codebase Intelligence: ingest repos, embed in Endee, answer with Groq."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "endee>=0.1.22",
    "fastapi>=0.135.1",
    "groq>=1.1.1",
    "inngest>=0.5.18",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
    "uvicorn>=0.42.0",
]

[project.scripts]
codebase-api = "main:main"

[tool.setuptools]
py-modules = [
    "main",
    "ast_parser",
    "chunks",
    "config",
    "embeddings",
    "endee_store",
    "ingestion",
    "llm",
    "retrieval",
    "streamlit_app",
    "workflow",
]
